Save to Excel
Save a dataframe to its own sheet in Excel. Don't include the index because that is just the row number.

Probably won't continue to save this particular dataframe to its own sheet, but useful for now for testing and this is the pattern to use for everything that will be saved.