Split puid dataframe by if match
If there is not a match to NARA, move to a separate dataframe so it can be tested for a match in another way.

If there was a match, remove the extra PRONOM column (comes from both FITS and NARA since they don't use the same name) and add a column for the match type.