Add columns for comparison
To increase the number of matches for name and for extension, making lowercase version for all and adding version to FITs format names. Haven't updated the merges yet to use them.

Also using a list of columns to filter df_nara, since making a new df has been causing problems for the PUID match, resulting in duplicated columns.