Option to use existing risk data
Save the dataframe that matches FITS format information to NARA risk, technical appraisal criteria, and other risk criteria to a CSV. Then, if this CSV is present from a previous iteration of the script, use that CSV for the analysis. This lets the user manually adjust risk rankings and generate summaries from the new data to account for imperfect automated matching.